Removed AudioCore from the server (still need to move the files). Removed the server's player. Fixed the team syncing. Fixed AI spawning. Fixed smallcar/truck spawning.
I've noticed the client crashes when a game starts (>= 5 players), I'm looking into it.
Added printf-like functionality to GameCore::mGui->outputToConsole on the server side - it takes a string and any number of arguments and parses them into a single string it prints to the console (why Microsoft couldn't do this for OutputDebugString I don't know).