Considerably improved the server's scheduler: unlocked the server now runs at 420,000 FPS, dropping to 5,000 FPS with 90 AI players. I have now separated the server's GUI rendering from the server's state updating so that they can run at different frame rates (both defined in ServerGraphics.cpp (defining in the header causes a full recompile each time they are changed)). Set either to 0 to unlock the frame rate. The graphics FPS will run no faster than the server FPS.
Locking the server FPS (to say 100) causes the CPU usage to drop to almost nothing, adding 90 AI players raises it to 60-80% of a core (it fluctuates), though the server FPS will not drop below 100 (unless the core maxes out, in which case it will as there are no more resources for it to eat into). In debug mode adding large numbers of AI players causes the graphics fps to drop to around 1. The only reason I can come up with for this is due to the huge amount of strings the debug server outputs (for collisions which are numerous for large numbers of players). Why this slows the graphics performance I have no idea, the graphics fps is ONLY for rendering the GUI now - no game work at all is done here.

The comments are currently out of date, though I very much doubt anyone would read them before I merge this so it doesn't matter.