Fixed the spawn screen so that it a) displays white preview images and the correct title text if it is a Free for All game. It also has the correct option selected at the end of the round (rather than always defaulting to the coupe).
Note: I am aware that if the next round is FFA you do NOT get white cars shown in the car selection screen. This is not due to me, but to the way the spawn screen is drawn. It is drawn at the end of the current round, NOT the start of the next, so you are basically getting your team colours for this round. If someone wants to fix it, ln227 in client/networking/PlayerPool.cpp needs moving to whether the hell the new round is synced.

Note: If the first round is FFA you still are asked to choose your team because you can not change your team during a game (so you are choosing the team for the other 2 rounds).